refactored the perf test
